Test minor change

+++ src/Main/SensorReader.cc
@@ -7,7 +7,7 @@
 long lastTime = millis();
 
 SensorReader::SensorReader(State &state) : state(state){
-	flmRpm = new Rpm{"rpmflm", 260, config.flmSensor};
+	flmRpm = new Rpm{"rpmflm", static_cast<uint8_t>(260), config.flmSensor};
 	frmRpm = new Rpm{"rpmfrm", 260, config.frmSensor};
 }
